+++ src/client/backends/ldp/impl/utils.ts
@@ -1,4 +1,3 @@
-import axios from 'axios';
 import {
   GraphQLBoolean,
   GraphQLField,
@@ -14,8 +13,9 @@ import {
   isNonNullType,
   isScalarType
 } from 'graphql';
-import { DataFactory, NamedNode, Parser, Quad, Store, Term } from 'n3';
-import { Graph, LdpClient, vocab } from '../../../../commons';
+import { DataFactory, NamedNode, Term } from 'n3';
+import { Graph, LdpClient } from '../../../../commons';
+import { getDirectivesMap } from '../../../../commons/util';
 import { RDFS } from '../../../../commons/vocab';
 
 const { namedNode } = DataFactory;
@@ -24,86 +24,6 @@ export type Primitive = boolean | number | string;
 export enum ResourceType {
   CONTAINER,
   DOCUMENT
-}
-
-/* eslint-disable @typescript-eslint/no-unused-vars */
-export async function getSubGraphArray(
-  source: Quad[],
-  className: string,
-  predicate: string | null,
-  args: Record<string, any>
-): Promise<Quad[][]> {
-  const store = new Store(source);
-  // TODO: generate subgraphs based on sub in [sub ?className ? ?]
-  const quadsOfQuads = store
-    .getSubjects(vocab.RDFS.a, namedNode(className), null)
-    .map(
-      async (sub) =>
-        await getSubGraph(source, className, predicate, { id: sub.value })
-    );
-  return Promise.all(quadsOfQuads);
-}
-/* eslint-enable @typescript-eslint/no-unused-vars */
-
-export async function getSubGraph(
-  source: Quad[],
-  className: string,
-  predicate: string | null,
-  args: Record<string, any>
-): Promise<Quad[]> {
-  const store = new Store(source);
-  const id = args?.id;
-  const topQuads = store
-    .getSubjects(vocab.RDFS.a, namedNode(className), null)
-    .flatMap((sub) => {
-      if (id) {
-        return sub.value === id ? store.getQuads(sub, null, null, null) : [];
-      } else {
-        return store.getQuads(sub, null, null, null);
-      }
-    });
-  const follow = (quads: Quad[], store: Store): Quad[] => {
-    return quads.reduce(
-      (acc, quad) =>
-        quad.object.termType === 'BlankNode' ||
-        quad.object.termType === 'NamedNode'
-          ? [
-              ...acc,
-              quad,
-              ...follow(store.getQuads(quad.object, null, null, null), store)
-            ]
-          : [...acc, quad],
-      [] as Quad[]
-    );
-  };
-  return follow(topQuads, store);
-}
-
-export async function getGraph(location: string): Promise<Quad[]> {
-  const doc = await axios.get(location);
-  return new Parser().parse(doc.data);
-}
-
-export function getDirectives(
-  type: GraphQLType | GraphQLField<any, any, any> | GraphQLInputField
-): Record<string, any> {
-  if (isListType(type)) {
-    return getDirectives(type.ofType);
-  }
-  if (isNonNullType(type)) {
-    return getDirectives(type.ofType);
-  }
-  return isScalarType(type) ? {} : type.extensions.directives ?? {};
-}
-
-export function getRawType(type: GraphQLType): GraphQLNamedType {
-  if (isListType(type)) {
-    return getRawType(type.ofType);
-  }
-  if (isNonNullType(type)) {
-    return getRawType(type.ofType);
-  }
-  return type;
 }
 
 export interface IntermediateResultInput {
@@ -147,12 +67,24 @@ export class IntermediateResult {
   }
 }
 
-export function getClassURI(type: GraphQLType): NamedNode {
-  return namedNode(getDirectives(type).is?.class ?? '');
+export function getRawType(type: GraphQLType): GraphQLNamedType {
+  if (isListType(type)) {
+    return getRawType(type.ofType);
+  }
+  if (isNonNullType(type)) {
+    return getRawType(type.ofType);
+  }
+  return type;
 }
 
-export function getPropertyIRI(type: GraphQLType): NamedNode {
-  return namedNode(getDirectives(type).property?.iri ?? '');
+export function getClassURI(type: GraphQLType): NamedNode {
+  return namedNode(getDirectivesMap(getRawType(type)).is?.class ?? '');
+}
+
+export function getPropertyIRI(
+  field: GraphQLField<any, any> | GraphQLInputField
+): NamedNode {
+  return namedNode(getDirectivesMap(field).property?.iri ?? '');
 }
 
 export function getCurrentDirective(
@@ -165,10 +97,8 @@ export function getCurrentDirective(
     const type = key
       ? (schema.getType(typename) as GraphQLObjectType).getFields()[key]!
       : schema.getType(typename);
-    return (
-      (type?.extensions?.directives as Record<string, Record<string, any>>) ??
-      {}
-    );
+
+    return getDirectivesMap(type!);
   }
   return {};
 }
@@ -190,7 +120,7 @@ export async function getInstanceById(
     );
   }
   const documentGraph = await ldpCLient.downloadDocumentGraph(documentUrl);
-  return documentGraph.find(namedNode(id), RDFS.a, classUri).map(
+  const result = documentGraph.find(namedNode(id), RDFS.a, classUri).map(
     (quad) =>
       new IntermediateResult({
         requestURL: targetUrl,
@@ -199,6 +129,7 @@ export async function getInstanceById(
         subject: quad.subject!
       })
   )[0];
+  return result;
 }
 
 export function convertScalarValue(