+++ test/execution.test.ts
@@ -111,6 +111,7 @@ describe('A GQL Schema can execute', () => {
     expect(result).not.toBeUndefined();
     expect(result.data).not.toBeUndefined();
     expect(result.data).toHaveProperty('contact');
+    console.log(result);
     const contact = (result.data! as any).contact;
     expect(contact).toMatchObject({
       id: 'http://example.org/cont/tdupont',